Major changes
Implemented a lightweight logging system
fixed several bugs about the generation of the linq file. all the
referenced assemblies are now loaded and included and referenced in the
linq file.
implemented few unit tests (namespace have been moved from the previous
bridge naming convention)
added the assemblyqualifiedname to the message being serialized from the
debugee to the debugger
changed info in the assemblyinfo
fixed a issue in the LINQBridgeExtension about the project reload